Partial fix for #8
The issue is in unk handling - unks aren't converted to romaji, even if
they're all kana. This fixes that. The issue of what to do with
non-Japanese input is more complicated - what do you do if your input
has cyrillic or hangul?

An unrelated issue is that cutlet seems to be putting out zenkaku
eiji sometimes, I need to check why that's happening.